fix: baseUrl이 안 됨 이슈
