+++ src/app/api/index.ts
@@ -43,7 +43,6 @@ const prefixUrl =
   process.env.NEXT_PUBLIC_APP_ENV === "production"
     ? process.env.NEXT_PUBLIC_HOST
     : process.env.NEXT_PRIVATE_HOST;
-
 export const kyJsonInstance = ky.create({
   prefixUrl,
   headers: {
