+++ src/view/signup/EmailVerification/index.tsx
@@ -21,6 +21,7 @@ const EmailVerification = () => {
   });
 
   const handleSubmit = (values: z.infer<typeof emailVerificationSchema>) => {
+    console.log("values.email: ", values.email);
     sendEmail(values.email);
   };
